Fix backup function that maked backups of symlinks

+++ install.sh
@@ -1,24 +1,24 @@
 #!/bin/bash
 d=~/.mydotfiles
-create_backup() {
+create_link() {
 	local filename=$(basename -- "$1")
 	local dirname=$(dirname "$1")
 	local backupname="${filename%.*}_old.${filename##*.}"
-	if [ -e "$1" ]; then
+	if [ -e "$1" ] && [[ ! -h "$1" ]]; then
 		mv "$1" "$dirname/$backupname"
 		echo -e "Backup of $filename was created"
+		ln -s "$d/linux/$filename" $pa
+		echo "$d/linux/$filename -> $pa"
+	else
+		echo -e "Backup of a symlink $pa was not created. No changes is done"
 	fi
 }
-#mv a_old.txt a.txt
-#create_backup "a.txt"
 
 loop_through() {
 	while read p; do
 		local filename=$(basename -- "$p")
 		eval local pa=$p
-		create_backup $pa
-		ln -s "$d/linux/$filename" $pa
-		echo "$d/linux/$filename -> $pa"
+		create_link $pa
 	done <"$d/linux.txt"
 }
